likita_proj/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.5.7
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.1.0
colorama==0.4.6
deprecation==2.1.0
dj-database-url==2.0.0
Django==4.2.1
django-bootstrap-v5==1.0.11
django-cors-headers==4.0.0
django-countries==7.5.1
django-environ==0.10.0
django-hvad==1.8.0
django-libs==2.0.3
Django-Verify-Email==2.0.3
djangorestframework==3.14.0
fontawesomefree==6.4.0
frozenlist==1.3.3
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
jsonschema==4.17.3
msgpack==1.0.5
multidict==6.0.4
openai==0.27.7
packaging==23.1
Pillow==9.5.0
psycopg2==2.9.6
PyJWT==2.7.0
pyrsistent==0.19.3
pytz==2023.3
redis==4.5.5
requests==2.31.0
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
tqdm==4.65.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.2
whitenoise==6.4.0
yarl==1.9.2
